[v0.9.0-beta.3] Add more modules and handle double pass music error
