+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "stroycord",
   "description": "La fameuse phrase \"je reviens\" de DevilKing a chaque depart de ce dernier de discord INCLUS MAINTENANT UN BOT MUSICAL",
-  "version": "0.9.0-beta.2",
+  "version": "0.9.0-beta.3",
   "engines": {
     "node": "16.14.*"
   },
